refactor(baseline): decompose god class into façade + 5 mixins
BaselineAnalyzer (1,535 LOC) → façade (~145 LOC) + 5 mixins:
- DataLoadingMixin: load_adherence_data, load_intervals_data, parse_skipped_replaced_sessions, _extract_reason, load_cardiovascular_coupling
- AnomalyDetectionMixin: detect_unsolicited_activities
- PatternAnalysisMixin: analyze_skip_reasons, analyze_day_of_week_patterns, analyze_workout_type_patterns
- MetricsMixin: validate_data_quality, calculate_adherence_metrics, calculate_tss_metrics, analyze_tsb_trajectory, calculate_cv_coupling_metrics
- ReportingMixin: generate_json_output, generate_markdown_report, _format_markdown_report

All 38 existing tests pass without modification (5/10 god scripts refactored).
