+++ magma_cycling/analysis/baseline/__init__.py
@@ -1,0 +1,1 @@
+"""BaselineAnalyzer mixin modules for god class decomposition."""
